glowguard-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.14.0
torch==2.1.1
torchvision==0.16.1
pymongo==4.6.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.1
pyjwt==2.8.1
email-validator==2.1.0
aiofiles==23.2.1
requests==2.31.0
pytest==7.4.3
